perf: upgrade pyright-internal 1.1.301 → 1.1.408 with performance fixes
Upgrades the vendored pyright-internal from 1.1.301 to 1.1.408 and
adapts pyright-scip for the new API (Uri instead of string paths,
ServiceProvider initialization, restructured type properties).

The new Pyright introduced memory management (cache eviction) designed
for long-running language servers that caused severe performance
regression in our batch indexer (17s → 70s). Three targeted fixes:

- Cache tokenizer output unconditionally (sourceFile.ts): new Pyright
  only cached tokenizer output for "open" files (LSP clients). Our
  batch indexer has no open files, so every file was re-tokenized on
  access during SCIP emission. This was the dominant cost (~45s).

- 10x cache eviction thresholds (program.ts): raised parsed file count
  threshold from 1000 to 10000 and type cache entry threshold from
  750K to 7.5M to prevent eviction during batch indexing workloads.

- Disable type checker for indexing (indexer.ts): the checker produces
  type cache entries that are never used during SCIP emission. Skipping
  it saves ~8.5s of analysis time with identical index output.

- Strip non-essential data after analysis (sourceFile.ts): clears
  diagnostics, symbol tables, and circular dependency tracking before
  emission to reduce memory pressure.

Also refactors main-impl.ts to extract runSingleThreaded for reuse.

+++ packages/pyright-internal/src/tests/samples/classes7.py
@@ -10,7 +10,8 @@ class BaseClass(Generic[T]):
     pass
 
 
-IntBaseClass = BaseClass[int]
+IntBaseClass = BaseClass[float]
+
 
 # This should generate an error because the same
 # base class is used twice.
